Refactor the Agent into go routines
This should be more or less the final architecture in terms
of execution contexts.
